Implement and spec new notification format "2"
Apple modified the Local and Push Notification Programming Guide last
April specifying a new format “2” and labeling simple and enhanced
formats as legacy.

New priority attribute. According to Apple it could be 10 or 5 (or
nothing?). Those values are not enforced, in case other values appear.

Expiration and identifier are no longer neccessary. The identifier is
added by the Houston::Client.

Using ‘n’ and ‘N‘ for packs and unpacks, since Apple documents that all
data is network order (big endian), and ‘i’ is native endianess.